requirements.txt
moviepy
Pillow
python-bidi
arabic-reshaper
yt-dlp
python-telegram-bot
python-telegram-bot[job-queue]
python-dotenv
pilmoji==2.0.4
emoji==1.7.0
requests